[CSA-35] Reverted the added class for code mutualisation to avoid to many parent class
