+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/HvdcRangeActionImpl.java
@@ -15,6 +15,7 @@ import com.farao_community.farao.data.crac_api.usage_rule.UsageRule;
 import com.powsybl.iidm.network.HvdcLine;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.extensions.HvdcAngleDroopActivePowerControl;
+import org.apache.commons.lang3.NotImplementedException;
 
 import java.util.Collections;
 import java.util.List;
@@ -28,14 +29,18 @@ import static com.farao_community.farao.commons.logs.FaraoLoggerProvider.TECHNIC
  * @author Godelaine de Montmorillon {@literal <godelaine.demontmorillon at rte-france.com>}
  * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
  */
-public class HvdcRangeActionImpl extends AbstractStandardRangeAction<HvdcRangeAction> implements HvdcRangeAction {
+public class HvdcRangeActionImpl extends AbstractRangeAction<HvdcRangeAction> implements HvdcRangeAction {
 
     private final NetworkElement networkElement;
+    private final List<StandardRange> ranges;
+    private final double initialSetpoint;
 
     HvdcRangeActionImpl(String id, String name, String operator, Set<UsageRule> usageRules, List<StandardRange> ranges,
                         double initialSetpoint, NetworkElement networkElement, String groupId, Integer speed) {
-        super(id, name, operator, usageRules, groupId, speed, ranges, initialSetpoint);
+        super(id, name, operator, usageRules, groupId, speed);
         this.networkElement = networkElement;
+        this.ranges = ranges;
+        this.initialSetpoint = initialSetpoint;
     }
 
     @Override
@@ -44,8 +49,60 @@ public class HvdcRangeActionImpl extends AbstractStandardRangeAction<HvdcRangeAc
     }
 
     @Override
+    public List<StandardRange> getRanges() {
+        return ranges;
+    }
+
+    @Override
     public Set<NetworkElement> getNetworkElements() {
         return Collections.singleton(networkElement);
+    }
+
+    @Override
+    public double getMinAdmissibleSetpoint(double previousInstantSetPoint) {
+        double minAdmissibleSetpoint = Double.NEGATIVE_INFINITY;
+        for (StandardRange range : ranges) {
+            switch (range.getRangeType()) {
+                case ABSOLUTE:
+                    minAdmissibleSetpoint = Math.max(minAdmissibleSetpoint, range.getMin());
+                    break;
+                case RELATIVE_TO_INITIAL_NETWORK:
+                    minAdmissibleSetpoint = Math.max(minAdmissibleSetpoint, initialSetpoint + range.getMin());
+                    break;
+                case RELATIVE_TO_PREVIOUS_INSTANT:
+                    minAdmissibleSetpoint = Math.max(minAdmissibleSetpoint, previousInstantSetPoint + range.getMin());
+                    break;
+                default:
+                    throw new NotImplementedException("Range Action type is not implemented yet.");
+            }
+        }
+        return minAdmissibleSetpoint;
+    }
+
+    @Override
+    public double getMaxAdmissibleSetpoint(double previousInstantSetPoint) {
+        double maxAdmissibleSetpoint = Double.POSITIVE_INFINITY;
+        for (StandardRange range : ranges) {
+            switch (range.getRangeType()) {
+                case ABSOLUTE:
+                    maxAdmissibleSetpoint = Math.min(maxAdmissibleSetpoint, range.getMax());
+                    break;
+                case RELATIVE_TO_INITIAL_NETWORK:
+                    maxAdmissibleSetpoint = Math.min(maxAdmissibleSetpoint, initialSetpoint + range.getMax());
+                    break;
+                case RELATIVE_TO_PREVIOUS_INSTANT:
+                    maxAdmissibleSetpoint = Math.min(maxAdmissibleSetpoint, previousInstantSetPoint + range.getMax());
+                    break;
+                default:
+                    throw new NotImplementedException("Range Action type is not implemented yet.");
+            }
+        }
+        return maxAdmissibleSetpoint;
+    }
+
+    @Override
+    public double getInitialSetpoint() {
+        return initialSetpoint;
     }
 
     @Override
@@ -101,13 +158,17 @@ public class HvdcRangeActionImpl extends AbstractStandardRangeAction<HvdcRangeAc
             return false;
         }
 
-        return this.networkElement.equals(((HvdcRangeAction) o).getNetworkElement());
+        return this.networkElement.equals(((HvdcRangeAction) o).getNetworkElement())
+                && this.ranges.equals(((HvdcRangeAction) o).getRanges());
 
     }
 
     @Override
     public int hashCode() {
         int hashCode = super.hashCode();
+        for (StandardRange range : ranges) {
+            hashCode += 31 * range.hashCode();
+        }
         hashCode += 31 * networkElement.hashCode();
         return hashCode;
     }
